Point
Println:Println formats using the default formats for its operands and writes to standard output. Spaces are always added between operands and a newline is appended. It returns the number of bytes written and any write error encountered.

Printf:Printf formats according to a format specifier and writes to standard output. It returns the number of bytes written and any write error encountered.

--Source https://golang.org/pkg/fmt/#example_Printf